requirements.txt
discord.py
yt-dlp
python-dotenv
nest_asyncio
certifi